packages/agent-workflows/pyproject.toml
[project]
name = "siili-agent-workflows"
version = "0.1.0"
description = "YAML-driven workflow engine for AI agents"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Siili Solutions Oyj", email = "info@siili.com"}
]
keywords = ["ai", "agents", "workflows", "yaml", "automation", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "python-dotenv>=1.0",
    "platformdirs>=4.0",
    "siili-ai-sdk>=0.2.0",
    "siili-coding-agents>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "ruff",
    "ty",
]

[project.urls]
Homepage = "https://github.com/siilisolutions/siili-ai-sdk"
Repository = "https://github.com/siilisolutions/siili-ai-sdk"

[project.scripts]
siili-agent-workflows = "agent_workflows.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["agent_workflows"]

[tool.hatch.build.targets.wheel.sources]
"agent_workflows" = "agent_workflows"

[tool.hatch.build]
include = [
    "agent_workflows/**/*.py",
    "agent_workflows/**/*.yml",
    "agent_workflows/**/*.yaml",
]

[tool.ruff]
line-length = 100
fix = true

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["W293"]

[tool.ruff.format]
quote-style = "double"

[tool.uv.sources]
siili-ai-sdk = { path = "../agent-sdk", editable = true }
siili-coding-agents = { path = "../coding-agents", editable = true }
